% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_col.R
\name{add_group_col}
\alias{add_group_col}
\title{Add a new variable based on some condition and group rows based on the new variable}
\usage{
add_group_col(data, name, where, value, add = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{name}{the name of the new variable to be created}

\item{where}{the condition in which the new variable should have a value (NA otherwise)}

\item{value}{an expression for the content of the new variable when the condition is true}

\item{add}{when `add = TRUE` (default), the new variable will add to existing groups. Set to `FALSE` to override all existing groups (passed to `dplyr::group_by`)}
}
\description{
Add a new variable based on some condition and group rows based on the new variable
}
\examples{
lexicon_str <-
'\\\\lx bureau
\\\\sn 1
\\\\de desk
\\\\sn 2
\\\\de office

\\\\lx langue
\\\\sn 1
\\\\de language
\\\\sn 2
\\\\de tongue
'

read_lexicon(file = lexicon_str, regex = "\\\\\\\\?([a-z]*)\\\\s?(.*)", into = c("code", "value")) \%>\%
    add_group_col(name = lx_group, where = code == "lx", value = paste0(line, ": ", value)) \%>\%
    add_group_col(name = sense_no, where = code == "sn", value = value)

}
